# Implement LogSubscriber and Notifications