--- a/lib/view_component/log_subscriber.rb
+++ b/lib/view_component/log_subscriber.rb
@@ -5,20 +5,28 @@ module ViewComponent
     def render_template(event)
       debug do
         message = +"  Rendered #{event.payload[:component_name]}"
-        message << " (Duration: #{event.duration.round(1)}ms | Allocations: #{event.allocations})"
+        message << " (#{duration_and_allocations(event)})"
       end
     end
 
     def render_collection(event)
       debug do
         message = +"  Rendered collection of #{event.payload[:component_name]}"
-        message << " [#{event.payload[:count]} times] (Duration: #{event.duration.round(1)}ms | Allocations: #{event.allocations})"
+        message << " [#{event.payload[:count]} times] (#{duration_and_allocations(event)})"
         message
       end
     end
 
     def logger
       ActionView::Base.logger
+    end
+
+    private
+
+    def duration_and_allocations(event)
+      message = +"Duration: #{event.duration.round(1)}ms"
+      message << " | Allocations: #{event.allocations}" if event.respond_to?(:allocations)
+      message
     end
   end
 end