--- a/test/view_component/log_subscriber_test.rb
+++ b/test/view_component/log_subscriber_test.rb
@@ -1,9 +1,7 @@
 # frozen_string_literal: true
 
-# require "abstract_unit"
 require "active_support/log_subscriber/test_helper"
 require "action_view/log_subscriber"
-# require "controller/fake_models"
 
 require "test_helper"