<!-- See https://github.com/github/view_component/blob/master/CONTRIBUTING.md#submitting-a-pull-request  -->

### Summary

Draft for #427.

This PR implement `ActiveSupport::Notifications` and `ActiveSupport::LogSubscriber`.

<!-- Provide a general description of the code changes in your pull
request... were there any bugs you had fixed? If so, mention them. If
these bugs have open GitHub issues, be sure to tag them here as well,
to keep the conversation linked together. -->

### Other Information

#### To Do
- [ ] Add tests
- [ ] Add more instrumentation?
- [ ] Confirm instrumentation names

#### Example of logs

```
web       | Started GET "/welcome/show" for 127.0.0.1 at 2020-08-02 15:51:43 +0200
web       | Processing by WelcomeController#show as HTML
web       |   Rendering welcome/show.html.erb within layouts/application
web       |   Rendered Elements::ButtonComponent (Duration: 7.9ms | Allocations: 2194)
web       |   Rendered Elements::ButtonComponent (Duration: 1.1ms | Allocations: 82)
web       |   Rendered welcome/show.html.erb within layouts/application (Duration: 13.2ms | Allocations: 2721)
web       |   Rendered Elements::HeaderComponent (Duration: 5.3ms | Allocations: 1828)
web       |   Rendered Elements::AlertComponent (Duration: 7.1ms | Allocations: 2970)
web       |   Rendered Elements::AlertComponent (Duration: 0.2ms | Allocations: 113)
web       |   Rendered Elements::FooterComponent (Duration: 5.6ms | Allocations: 1870)
web       |   Rendered Elements::ContainerComponent (Duration: 23.3ms | Allocations: 7050)
web       | Completed 200 OK in 45ms (Views: 43.2ms | ActiveRecord: 0.0ms | Allocations: 11721)
```

#### Questions

- What does `!` mean on beginning of instrumentation name ([example](https://github.com/rails/rails/blob/528b62e3869288b5f3ffbd1491f10bad00958c08/actionview/lib/action_view/template.rb#L384))? (answer: https://twitter.com/rafaelfranca/status/1289978377550442496)
- Does `render_template` instrumentation should include compilation? ([`view_component` does](https://github.com/rails/rails/blob/528b62e3869288b5f3ffbd1491f10bad00958c08/actionview/lib/action_view/template.rb#L180-L183))

<!-- If there's anything else that's important and relevant to your pull
request, mention that information here. This could include
benchmarks, or other information.

If you are updating any of the CHANGELOG files or are asked to update the
CHANGELOG files by reviewers, please add the CHANGELOG entry at the top of the file. -->