@joelhawksley 2020-08-03
@nicolas-brousse fantastic! Thank you SO MUCH for opening a PR.

I'm curious: what is the goal with having render times for each component? I only ask as we generally haven't built `view_component` to optimize for speed in local development. In fact, we've focused on speed in production, often to the detriment of speed in local development.

That aside, I do think that having the allocations for rendering each component is quite valuable either way ❤️ , and that we should proceed with this work, even if we just limit it to reporting allocations for now.