@nicolas-brousse 2020-08-17
@joelhawksley I'm back from holidays, I'll have some time this week to continue this PR :slightly_smiling_face: 

> I'm curious: what is the goal with having render times for each component? I only ask as we generally haven't built `view_component` to optimize for speed in local development. In fact, we've focused on speed in production, often to the detriment of speed in local development.

For now I simply take inspiration from what `action_view` does, to keep the same format. But we could remove the render times if it is not relevant.