@joelhawksley 2020-08-17
@nicolas-brousse I think removing the compile time from the calculation is sufficient.