@joelhawksley 2020-11-23
@nicolas-brousse thanks for cleaning this up! Mind adding a CHANGELOG entry and some docs to `docs/index.md`?