@nicolas-brousse 2020-11-30
@joelhawksley I updated the CHANGELOG, but not sure what to specify in `docs/index.md`.