@BlakeWilliams 2020-12-04
👋 I really like this change, but I ran benchmarks on `master` and it looks like we lost our performance wins over `partials` after this was merged. 😞 

I ran `bundle exec rake benchmark` against both and got these results:


Before:
```
Comparison:
          component::   182965.1 i/s
            partial::    17118.3 i/s - 10.69x  (± 0.00) slower
```

After:
```
Comparison:
          component::    18380.1 i/s
            partial::    17635.6 i/s - same-ish: difference falls within error
```

I plan on looking at this and seeing if there's some kind of middle ground or other wins here. Like maybe running the subscriber only in development.