@nicolas-brousse 2020-12-04
Thanks for your feedback, and sorry for the performance issues.

No worries I totally understand your revert.

I'll try to found a bit of time later and check if I saw an other implementation that does not touch performances.